requirements.txt
langchain
openai
boto3
python-dotenv
fastapi
uvicorn
botbuilder-core
botbuilder-schema
aiohttp

